\name{mypkg-package}


\alias{mypkg-package}
\alias{my-pkg}

\docType{package}

\title{
  epiplus
}

\description{
My package description
}

\details{
  \tabular{ll}{
    Package: \tab  my-pkg \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.1 \cr
    Date: \tab  01/01/2000 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Other packages by the same authors :

  \code{\link{lazyMe}}, \code{\link{dataDesc}}, \code{\link{manageR}},
  \code{\link{surveyscores}}, \code{\link{sampleSize}}, \code{\link{epiplus}}
}

\examples{

}
